refactor(issue-12): migrate live tests to BexioApiNet.E2eTests
Move 9 live test files from src/BexioApiNet.Tests/Tests/ to
tests/BexioApiNet.E2eTests/Tests/, updating namespaces from
BexioApiNet.Tests.Tests.* to BexioApiNet.E2eTests.Tests.* and changing
the base class from TestBase to BexioE2eTestBase. Relocate the Assets/
PDFs and wire them into the new .csproj. Delete the old test project.
