--- src/BexioApiNet.Tests/Tests/Banking/BankAccount/GetAll.cs
@@ -1,50 +1,0 @@
-﻿/*
-MIT License
-
-Copyright (c) 2022 Philip Näf <philip.naef@amanda-technology.ch>
-Copyright (c) 2022 Manuel Gysin <manuel.gysin@amanda-technology.ch>
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
-namespace BexioApiNet.Tests.Tests.Banking.BankAccount;
-
-/// <summary>
-///
-/// </summary>
-public class TestGetAll : TestBase
-{
-    /// <summary>
-    ///
-    /// </summary>
-    [Test]
-    public async Task GetAll()
-    {
-        Assert.That(BexioApiClient, Is.Not.Null);
-
-        var res = await BexioApiClient!.BankingBankAccounts.Get();
-        Assert.That(res, Is.Not.Null);
-        Assert.Multiple(() =>
-        {
-            Assert.That(res.IsSuccess, Is.True);
-            Assert.That(res.ApiError, Is.Null);
-            Assert.That(res.Data?.First().Name, Is.Not.Null);
-        });
-    }
-}
