+++ tests/BexioApiNet.E2eTests/Assets/letter.pdf
