+++ tests/BexioApiNet.E2eTests/Assets/letter2.pdf
