+++ tests/BexioApiNet.E2eTests/Tests/Banking/BankAccount/GetAll.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
 MIT License
 
 Copyright (c) 2022 Philip Näf <philip.naef@amanda-technology.ch>
@@ -23,12 +23,12 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-namespace BexioApiNet.Tests.Tests.Accounting.Taxes;
+namespace BexioApiNet.E2eTests.Tests.Banking.BankAccount;
 
 /// <summary>
 ///
 /// </summary>
-public class TestGetAll : TestBase
+public class TestGetAll : BexioE2eTestBase
 {
     /// <summary>
     ///
@@ -38,14 +38,13 @@ public class TestGetAll : TestBase
     {
         Assert.That(BexioApiClient, Is.Not.Null);
 
-        var res = await BexioApiClient!.Taxes.Get(autoPage: true);
-
+        var res = await BexioApiClient!.BankingBankAccounts.Get();
         Assert.That(res, Is.Not.Null);
         Assert.Multiple(() =>
         {
             Assert.That(res.IsSuccess, Is.True);
             Assert.That(res.ApiError, Is.Null);
-            Assert.That(res.Data?.First().Id, Is.Not.Null);
+            Assert.That(res.Data?.First().Name, Is.Not.Null);
         });
     }
 }
